+++ app/Http/Controllers/Api/ConversionController.php
@@ -583,7 +583,8 @@ class ConversionController extends ApiController
      * Get elastic search data
      *
      * @param string api_key - required
-     * @param string es_id - required
+     * @param string resource_id - required
+     * @param string version - optional
      *
      * @return json with json data or error
      */
@@ -591,11 +592,21 @@ class ConversionController extends ApiController
     {
         $post = $request->all();
 
-        $validator = \Validator::make($post, ['es_id' => 'required|int|exists:elastic_data_set,id']);
+        $validator = \Validator::make($post, [
+            'resource_id'   => 'required|int|exists:resources,id',
+            'version'       => 'sometimes|int|exists:elastic_data_set,version',
+        ]);
 
         if (!$validator->fails()) {
+            if (!isset($post['version'])) {
+                $resource = Resource::find($post['resource_id']);
+                $version = $resource->version;
+            } else {
+                $version = $post['version'];
+            }
+
             try {
-                $data = ElasticDataSet::getElasticData($post['es_id']);
+                $data = ElasticDataSet::getElasticData($post['resource_id'], $version);
 
                 return $this->successResponse($data);
             } catch (\Exception $ex) {
@@ -610,7 +621,8 @@ class ConversionController extends ApiController
      * Get elastic search data
      *
      * @param string api_key - required
-     * @param string es_id - required
+     * @param string resource_id - required
+     * @param string version - optional
      *
      * @return json with xml data or error
      */
@@ -618,11 +630,21 @@ class ConversionController extends ApiController
     {
         $post = $request->all();
 
-        $validator = \Validator::make($post, ['es_id' => 'required|int|exists:elastic_data_set,id']);
+        $validator = \Validator::make($post, [
+            'resource_id'   => 'required|int|exists:resources,id',
+            'version'       => 'sometimes|int|exists:elastic_data_set,version',
+        ]);
 
         if (!$validator->fails()) {
+            if (!isset($post['version'])) {
+                $resource = Resource::find($post['resource_id']);
+                $version = $resource->version;
+            } else {
+                $version = $post['version'];
+            }
+
             try {
-                $data = ElasticDataSet::getElasticData($post['es_id']);
+                $data = ElasticDataSet::getElasticData($post['resource_id'], $version);
                 $data = $this->getXML($data);
 
                 return $this->successResponse($data);
@@ -654,7 +676,8 @@ class ConversionController extends ApiController
      * Get ellastic search data
      *
      * @param string api_key - required
-     * @param string es_id - required
+     * @param string resource_id - required
+     * @param string version - optional
      *
      * @return json with csv data or error
      */
@@ -662,11 +685,21 @@ class ConversionController extends ApiController
     {
         $post = $request->all();
 
-        $validator = \Validator::make($post, ['es_id' => 'required|int|exists:elastic_data_set,id']);
+        $validator = \Validator::make($post, [
+            'resource_id'   => 'required|int|exists:resources,id',
+            'version'       => 'sometimes|int|exists:elastic_data_set,version',
+        ]);
 
         if (!$validator->fails()) {
+            if (!isset($post['version'])) {
+                $resource = Resource::find($post['resource_id']);
+                $version = $resource->version;
+            } else {
+                $version = $post['version'];
+            }
+
             try {
-                $data = ElasticDataSet::getElasticData($post['es_id']);
+                $data = ElasticDataSet::getElasticData($post['resource_id'], $version);
                 $data = $this->getCSV($data);
 
                 return $this->successResponse($data);
@@ -844,7 +877,8 @@ class ConversionController extends ApiController
      * Get elastic search data
      *
      * @param string api_key - required
-     * @param string es_id - required
+     * @param string resource_id - required
+     * @param string version - optional
      *
      * @return json with kml data or error
      */
@@ -852,11 +886,21 @@ class ConversionController extends ApiController
     {
         $post = $request->all();
 
-        $validator = \Validator::make($post, ['es_id' => 'required|int|exists:elastic_data_set,id']);
+        $validator = \Validator::make($post, [
+            'resource_id'   => 'required|int|exists:resources,id',
+            'version'       => 'sometimes|int|exists:elastic_data_set,version',
+        ]);
 
         if (!$validator->fails()) {
+            if (!isset($post['version'])) {
+                $resource = Resource::find($post['resource_id']);
+                $version = $resource->version;
+            } else {
+                $version = $post['version'];
+            }
+
             try {
-                $data = ElasticDataSet::getElasticData($post['es_id']);
+                $data = ElasticDataSet::getElasticData($post['resource_id'], $version);
                 $data = $this->getKML($data);
 
                 return $this->successResponse($data);
@@ -887,7 +931,8 @@ class ConversionController extends ApiController
      * Get elastic search data
      *
      * @param string api_key - required
-     * @param string es_id - required
+     * @param string resource_id - required
+     * @param string version - optional
      *
      * @return json with rdf data or error
      */
@@ -895,11 +940,21 @@ class ConversionController extends ApiController
     {
         $post = $request->all();
 
-        $validator = \Validator::make($post, ['es_id' => 'required|int|exists:elastic_data_set,id']);
+        $validator = \Validator::make($post, [
+            'resource_id'   => 'required|int|exists:resources,id',
+            'version'       => 'sometimes|int|exists:elastic_data_set,version',
+        ]);
 
         if (!$validator->fails()) {
+            if (!isset($post['version'])) {
+                $resource = Resource::find($post['resource_id']);
+                $version = $resource->version;
+            } else {
+                $version = $post['version'];
+            }
+
             try {
-                $data = ElasticDataSet::getElasticData($post['es_id']);
+                $data = ElasticDataSet::getElasticData($post['resource_id'], $version);
                 $data = $this->getRDF($data);
 
                 return $this->successResponse($data);