+++ app/Resource.php
@@ -27,13 +27,12 @@ class Resource extends Model implements TranslatableInterface
     const FORMAT_KML = 3;
     const FORMAT_RDF = 4;
     const FORMAT_XML = 5;
-    const FORMAT_WMS = 6;
-    const FORMAT_TSV = 7;
-    const FORMAT_XSD = 8;
-    const FORMAT_ODS = 9;
-    const FORMAT_SLK = 10;
-    const FORMAT_RTF = 11;
-    const FORMAT_ODT = 12;
+    const FORMAT_TSV = 6;
+    const FORMAT_XSD = 7;
+    const FORMAT_ODS = 8;
+    const FORMAT_SLK = 9;
+    const FORMAT_RTF = 10;
+    const FORMAT_ODT = 11;
 
     const HTTP_POST = 1;
     const HTTP_GET = 2;
@@ -74,8 +73,6 @@ class Resource extends Model implements TranslatableInterface
                 return self::FORMAT_KML;
             case 'RDF':
                 return self::FORMAT_RDF;
-            case 'WMS':
-                return self::FORMAT_WMS;
             case 'XML':
                 return self::FORMAT_XML;
             case 'XSD':
@@ -94,7 +91,6 @@ class Resource extends Model implements TranslatableInterface
             self::FORMAT_JSON   => 'JSON',
             self::FORMAT_KML    => 'KML',
             self::FORMAT_RDF    => 'RDF',
-            self::FORMAT_WMS    => 'WMS',
             self::FORMAT_XML    => 'XML',
             self::FORMAT_TSV    => 'TSV',
             self::FORMAT_XSD    => 'XSD',
@@ -150,7 +146,7 @@ class Resource extends Model implements TranslatableInterface
     public static function getAllowedFormats()
     {
         return [
-            'CSV', 'JSON', 'KML','RDF','WMS','XML','XLSX', 'XLS', 'TXT', 'TSV', 'XSD', 'ODS', 'SLK', 'RTF', 'ODT'
+            'CSV', 'JSON', 'KML', 'RDF', 'XML', 'XLSX', 'XLS', 'TXT', 'TSV', 'XSD', 'ODS', 'SLK', 'RTF', 'ODT'
         ];
     }
 }