+++ CHANGELOG.md
@@ -31,6 +31,7 @@ Security - in case of vulnerabilities.
 - Visual glitches in IE and mobile version
 - Bugs in showing system information about records
 - Dubious column names in documents and organisations
+- Resource download
 
 ## [0.6.2] - 2018-10-17
 ### Fixed